controller: moved parse_duration to utils.rs
